tools/requirements-watermark.txt
Pillow~=11.0
PyYAML~=6.0
pillow-heif~=0.22
